requirement.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
amqp==5.2.0
anyio==4.2.0
attrs==23.2.0
backoff==2.2.1
billiard==4.2.0
blinker==1.7.0
boto3==1.26.122
botocore==1.29.122
cachetools==5.3.2
celery==5.3.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dataclasses-json==0.6.4
distro==1.9.0
dnspython==2.6.1
Flask==2.3.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
importlib-metadata==7.0.1
invoke==2.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.6
langfuse==1.0.18
langsmith==0.0.87
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.12.0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pinecone-client==2.2.2
prompt-toolkit==3.0.43
protobuf==4.25.3
pyarrow==15.0.0
pydantic==1.10.14
pydeck==0.8.1b0
Pygments==2.17.2
pypdf==3.15.4
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.4
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
s3transfer==0.6.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.27
streamlit==1.31.1
streamlit-chat==0.1.1
tenacity==8.2.3
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
uuid==1.30
validators==0.22.0
vine==5.1.0
watchdog==3.0.0
wcwidth==0.2.13
Werkzeug==2.3.1
yarl==1.9.4
zipp==3.17.0
